+++ src/components/pagination/PagingBar.js
