+++ src/layouts/Layout.js
@@ -1,13 +1,25 @@
-import Header from "./Header";
-import Footer from "./Footer";
-import Main from "../components/common/Main/Main";
+import { Outlet } from "react-router-dom";
+import { ThemeProvider } from "react-bootstrap";
+import Header from "../components/common/Header";
+import Footer from "../components/common/Footer";
+import Wrapper from "../components/common/Wrapper";
+
 function Layout() {
 
     return (
         <>
-            <Header/>
-            <Main/>
-            <Footer/>
+            <ThemeProvider
+                breakpoints={[ 'xxxl', 'xxl', 'xl', 'lg', 'md', 'sm', 'xs', 'xxs' ]}
+                minBreakpoint="xs"
+            >
+                <Header />
+                <main className="main">
+                    <Wrapper>
+                        <Outlet />
+                    </Wrapper>
+                </main>
+                <Footer />
+            </ThemeProvider>
         </>
     );
 }
